Move utils file to somewhere more upgradable and extensible
