+++ packages/template-retail-react-app/app/components/_app/index.jsx
@@ -85,7 +85,7 @@ import {
 import Seo from '@salesforce/retail-react-app/app/components/seo'
 import ShopperAgent from '@salesforce/retail-react-app/app/components/shopper-agent'
 import {getPathWithLocale} from '@salesforce/retail-react-app/app/utils/url'
-import {parseCommerceAgentSettings} from '@salesforce/retail-react-app/config/utils'
+import {parseCommerceAgentSettings} from '@salesforce/retail-react-app/app/utils/config-utils'
 
 const PlaceholderComponent = () => (
     <Center p="2">
